python/requirements.txt
PyJWT
cryptography
bcrypt
flask
flask_jwt_extended